Update __init__.py
The use of only 20 elements to determine the dtype is creating problems and in some cases might lead to undetected data corruption. As a temporary measure I've changed the code to consider all elements rather than just 20 random elements. For efficiency we could consider an alternative method (we discussed heuristics based on the sum previously), but it'd be good to additionally (or instead) give users the option to specify dtypes when loading in the data (a possible use is if different data sets are read in separately and are to be combined later -- even when all elements in a given data set are considered for the determination of dtype, a missmatch between data sets can still occur).